kubernetes/jupyterhub: disable user scheduler pods
The "user scheduler" in Jupyterhub is a feature which schedules user
pods on the most utilized node in the cluster. While this plays well in
environments which have node autoscaling enabled, the data-infra-apps
cluster does not have such autoscaling configured. This consequently
causes node over-utilization and an eventual failure to schedule
altogether once the node becomes full.