fix(ci/service-build-release): parse BUILD_GIT_TAG for default BUILD_APP and BUILD_ID
The previous method of embedding build metadata directly in the tag
annotation meant there was never a need to parse the tag nam in order to
load it. Now that build metadata is stored in git notes, the tag should
be relied on to provide the metadata encoded in its name for the
initial build (as the git note will not exist prior to the build). This
reliance means it is also now a requirement to validate the tag form in
order to ensure it can be properly parsed.

The parsing method is chosen rather than requiring the push of a
pre-populated git note so that the burden of managing git note metadata
can be borne exclusively by the pipeline and need not be passed along to
the operator.