fix(github/service-build): fix variable evaluation
Setting BUILD_GIT_TAG via git-describe is, for inexplicable reasons, not
possible within an actions environment. It is furthermore unnecessary
due to the availability of GITHUB_REF.

GITHUB_JOB appears to be useless outside of an actual running step due
to the fact that it is not set until the job begins running.