ci: implement diff-tree powered releases
This changeset introduces the capability for pipelines to internally
compute which applications to deploy in a given run by performing a
diff-tree between a previous deploy point and HEAD.

There are two included methods for selecting a previous deploy point:

1. Use the first parent (base) of a merge commit (only useful when HEAD
 is a merge commit)
2. Use the commit pointed to by the most recent note in a deployment
 notes ref

This release method is better suited to a multi-application repo. The
previous method of assuming only one application release per pipeline-
run does not cover the perfectly valid scenario of handling updates to
multiple applications at once.