charts/metabase: inject default MB_SITE_URL from ingress
Metabase will use the internal container IP by default to form its
configured site url. This causes issues such as users receiving broken
links in account activation emails. This changeset adds the capability
to the chart to set the site url to whatever hostname is configured in
the ingress configuration.