Add test for table emptiness
This test is useful for validation of general query. The way to use it 
is to create a view that selects some invalid records. If that view 
materializes to an empty table (i.e., there are no invalid records) then 
the test passes. If the view materializes some non-empty records (i.e., 
there are invalid records), then the test fails.

Since the testing framework expects a list of fields, use the 
check_empty test like:

tests:
  check_empty:
    - "*"