gtfs_rt_archive/threads/writer: refactor GCPBucketWriter for async write
The BaseWriter is designed to be single-threaded for backends where
multiple writers may cause contention or corruption (e.g., appending to
a single file), but this concern does not apply to GCP buckets. Making
writes async will lower the total time to complete all txns on each tick
since it eliminates the need for sequential completion of writes.