service-release-candidate.yml: base all candidate ancestry on prod
Using the previous HEAD of the candidate branch to base subsequent
candidate projections on resulted in a git history which did not have
the desired effect of showing the minimum amount of changes between
candidate releases. This change has the added benefit of always showing
how a change will be applied on prod via its diff.