[images/jupyter-singleuser] Update high criticality dependencies (#4148)
* rm calitp-map-utils

There were version conflicts between calitp-map-utils and other package
upgrades needed for jupyter0singleuser. calitp-map-utils does not appear
to be used in data-analyses, csis-metrics or other analyst repos. If it
is needed in this image, we would just need to update dependencies in
calitp-map-utils to be compatible.

* Upgrade ydata-profiling

This package <4.9 had a few high criticality dependabot warnings
- https://github.com/cal-itp/data-infra/security/dependabot/1929
- https://github.com/cal-itp/data-infra/security/dependabot/1927
- https://github.com/cal-itp/data-infra/security/dependabot/1925
- also updates matplotlib a few minor versions for macos compatibility

* Upgrade tornado due to dependabot high criticality

https://github.com/cal-itp/data-infra/security/dependabot/2216

* rm/unpin setuptools transitive dependency

This should hopefully resolve the high criticality dependabot warnings
for <78.1.1
- https://github.com/cal-itp/data-infra/security/dependabot/2208
- https://github.com/cal-itp/data-infra/security/dependabot/2082

* Upgrade protobuf due to dependabot high criticality

Upgrade by pinning to patch version to 5.29.5 https://github.com/cal-itp/data-infra/security/dependabot/2235

* Bump jupyter-singleuser image version to 2025.8.1

* Add new profile image for testing